post product fix

+++ src/services/deleteKey.service.js
@@ -2,16 +2,26 @@ const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 
 const deleteKeys = async (keyId) => {
-
-     for(let id of keyId){
-        console.log(id)
-        const keysToDeleted = await prisma.key.delete({
-            where: {
-                id: Number(id)
+    // console.log(keyId)
+    if(keyId.length >= 1){
+        for(let id of keyId){
+            const searchKeyToDelete = await prisma.key.findUnique({
+                where: {
+                    id: Number(id)
+                } 
+            })
+            if(searchKeyToDelete) {
+                const keysToDeleted = await prisma.key.delete({
+                    where: {
+                        id: Number(id)
+                    }
+                })
+                console.log(keysToDeleted)
             }
-        })
-        console.log(keysToDeleted)
-    }    
+        }
+    } else {
+        console.log('Key doesnt exists')
+    } 
 }
 
 module.exports = {
